+++ src/GluttonyCombo/GluttonyCombo/Combos/PvE/ALL/ALL.cs
@@ -1,4 +1,4 @@
-using Dalamud.Game.ClientState.Objects.Types;
+﻿using Dalamud.Game.ClientState.Objects.Types;
 using ECommons.ExcelServices;
 using ECommons.GameHelpers;
 using System.Collections.Generic;
@@ -10,11 +10,12 @@ namespace GluttonyCombo.Combos.PvE;
 
 internal partial class All
 {
-    public const uint SavageBlade = 11, // Used to block user input.
+    public const uint 
         SingleTargetDPS = 1_000_000, // These next set are for native actions, if SE ever adds values up to these we would have to adjust but it's unlikely.
         AoEDPS = 1_000_001,
         SingleTargetHeals = 1_000_002,
         AoeHeals = 1_000_003,
+        Cease = 1_000_004, // Used as a replacement for Savage Blade blocking. Must stay below Items (2_000_000): several paths treat >= Items as an item id.
         Items = 2_000_000; // Ids on top of this will be for items, created dynamically
 
     public static class Buffs
@@ -101,7 +102,7 @@ internal partial class All
         protected override uint Invoke(uint actionID) =>
             actionID is RoleActions.Tank.Reprisal &&
             GetStatusEffectRemainingTime(RoleActions.Tank.Debuffs.Reprisal, CurrentTarget, true) > Config.AllTankReprisalThreshold
-                ? SavageBlade
+                ? Cease
                 : actionID;
     }
 
@@ -216,7 +217,7 @@ internal partial class All
         protected override uint Invoke(uint actionID) =>
             actionID is RoleActions.Caster.Addle &&
             GetStatusEffectRemainingTime(RoleActions.Caster.Debuffs.Addle, CurrentTarget, true) > Config.AllCasterAddleThreshold
-                ? SavageBlade
+                ? Cease
                 : actionID;
     }
 
@@ -264,7 +265,7 @@ internal partial class All
         protected override uint Invoke(uint actionID) =>
             actionID is RoleActions.Melee.Feint &&
             GetStatusEffectRemainingTime(RoleActions.Melee.Debuffs.Feint, CurrentTarget, true) > Config.AllMeleeFeintThreshold
-                ? SavageBlade
+                ? Cease
                 : actionID;
     }
 
@@ -274,7 +275,7 @@ internal partial class All
 
         protected override uint Invoke(uint actionID) =>
             actionID is RoleActions.Melee.TrueNorth && HasStatusEffect(RoleActions.Melee.Buffs.TrueNorth)
-                ? SavageBlade
+                ? Cease
                 : actionID;
     }
 
@@ -289,7 +290,7 @@ internal partial class All
              GetStatusEffectRemainingTime(MCH.Buffs.Tactician, anyOwner: true) > Config.AllRangedMitigationThreshold ||
              GetStatusEffectRemainingTime(DNC.Buffs.ShieldSamba, anyOwner: true) > Config.AllRangedMitigationThreshold) &&
             IsOffCooldown(actionID)
-                ? SavageBlade
+                ? Cease
                 : actionID;
     }