+++ src/GluttonyCombo/GluttonyCombo/CustomCombo/WrathOpener.cs
@@ -145,7 +145,7 @@ public abstract class WrathOpener
         get;
         set
         {
-            if (value != All.SavageBlade)
+            if (value != All.Cease)
                 field = value;
         }
     }
@@ -248,7 +248,7 @@ public abstract class WrathOpener
 
                 if (skipped)
                 {
-                    actionID = All.SavageBlade;
+                    actionID = All.Cease;
                     return true;
                 }
 
@@ -257,7 +257,7 @@ public abstract class WrathOpener
                 float startValue = (VeryDelayedWeaveSteps.Any(x => x == OpenerStep)) ? 1f : 1.25f;
                 if ((DelayedWeaveSteps.Any(x => x == OpenerStep) || VeryDelayedWeaveSteps.Any(x => x == OpenerStep)) && !CanDelayedWeave(startValue, 0) && RemainingGCD > 0)
                 {
-                    actionID = All.SavageBlade;
+                    actionID = All.Cease;
                     return true;
                 }
 
@@ -284,7 +284,7 @@ public abstract class WrathOpener
                     if ((DateTime.Now - DelayedAt).TotalSeconds < DelayedSecs && !PartyInCombat())
                     {
                         ActionWatching.TimeLastActionUsed = DateTime.Now; //Hacky workaround for TN jobs
-                        actionID = All.SavageBlade;
+                        actionID = All.Cease;
                         return true;
                     }
                 }